Conflicts:
	Glitch/Game1.cs

+++ Glitch/GameVariables.cs
@@ -30,9 +30,9 @@ namespace Glitch
         public static Vector2 PLAYER_DIMENSIONS;
         public static Vector2 TRAP_DIMENSIONS;
         public static Vector2 BULLET_DIMENSIONS;
-        public const float PLAYER_SCALE = 0.25f;
-        public const float ENEMY_SCALE = 0.1f;
-        public const float TRAP_SCALE = 1f;
-        public const float BULLET_SCALE = 0.03f;
+        public const double PLAYER_SCALE = 0.25;
+        public const double ENEMY_SCALE = 0.1;
+        public const double TRAP_SCALE = 1;
+        public const double BULLET_SCALE = 0.03;
     }
 }
